pyproject.toml
[project]
name = "streamlit-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.0",
    "plotly>=6.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.38",
    "streamlit>=1.42.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.6",
]

[tool.ruff]
line-length = 79